+++ threedi_modelchecker/checks/other.py
@@ -551,3 +551,34 @@ class PumpStorageTimestepCheck(BaseCheck):
 
     def description(self) -> str:
         return f"{self.column_name} will empty its storage faster than one timestep, which can cause simulation instabilities"
+
+
+class BetaColumnsCheck(BaseCheck):
+    """Check that no beta columns were used in the database"""
+
+    def get_invalid(self, session: Session) -> List[NamedTuple]:
+        return session.query(self.table).filter(self.column.isnot(None)).all()
+
+    def description(self) -> str:
+        return f"{self.column_name} is a beta feature, which is still under development; please do not use it yet."
+
+
+class BetaValuesCheck(BaseCheck):
+    """Check that no beta features were used in the database"""
+
+    def __init__(
+        self,
+        column,
+        values: list = [],
+        filters=None,
+        level=CheckLevel.ERROR,
+        error_code=0,
+    ):
+        super().__init__(column, filters, level, error_code)
+        self.values = values
+
+    def get_invalid(self, session: Session) -> List[NamedTuple]:
+        return session.query(self.table).filter(self.column.in_(self.values)).all()
+
+    def description(self) -> str:
+        return f"The value you have used for {self.column_name} is still in beta; please do not use it yet."