+++ CHANGES.rst
@@ -5,6 +5,11 @@ Changelog of threedi-modelchecker
 2.0.2 (unreleased)
 ------------------
 
+- Add support for designating beta features in threedi-schema. If a user puts a
+  non-null value in a column marked as beta in threedi-schema, a BetaFeaturesCheck
+  error 1300 will be raised by the modelchecker. The allow-beta flag has been added
+  to the CLI interface to disable this check temporarily.
+
 - Add errors and warnings for vegetation_drag input. Both rasters and global values.
 
 - Added check 73: groundwater boundaries are allowed only when there is
@@ -20,7 +25,7 @@ Changelog of threedi-modelchecker
 2.0.1 (2023-03-20)
 ------------------
 
-- Pin minor version for threede-schema dependency.
+- Pin minor version for threedi-schema dependency.
 
 
 2.0.0 (2023-03-20)